requirements.txt
# requirements.txt
numpy
pandas
pyyaml
matplotlib
streamlit
plotly
